SpecificIsolatedST cleanup
